✨ feat : 리뷰 작성

+++ src/main/java/org/likelion/_thon/silver_navi/domain/nursingfacility/web/dto/ReviewRes.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record ReviewRes(
         Long id,
         String authorName,
-        BigDecimal rating,
+        Integer rating,
         String content,
         LocalDateTime createdAt
 ) {
